Moved entities to Core, fixed references/dependecies
