+++ src/PlexBackend.Core/Entities/ProjectGroup.cs
@@ -4,14 +4,12 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace PlexBackend.Infrastructure.ContextModels
+namespace PlexBackend.Core.ContextModels
 {
-    public class Project_Term
+    public class ProjectGroup
     {
         public int Id { get; set; }
-        public int TermId { get; set; }
         public int ProjectId { get; set; }
-
-        public virtual Term Term { get; set; }
+        public string StudentPCNs { get; set; }
     }
 }
