+++ src/PlexBackend.Core/Interfaces/IStudentChoiceRepository.cs
@@ -1,11 +1,11 @@
-﻿using PlexBackend.Infrastructure.ContextModels;
+﻿using PlexBackend.Core.ContextModels;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace PlexBackend.Infrastructure.Interfaces
+namespace PlexBackend.Core.Interfaces
 {
     public interface IStudentChoiceRepository : IRepository<StudentChoice>
     {
