+++ src/PlexBackend.Core/PlexBackend.Core.csproj
@@ -4,8 +4,4 @@
         <TargetFramework>net5.0</TargetFramework>
     </PropertyGroup>
 
-    <ItemGroup>
-      <ProjectReference Include="..\PlexBackend.Infrastructure\PlexBackend.Infrastructure.csproj" />
-    </ItemGroup>
-
 </Project>
