+++ src/PlexBackend.Infrastructure/PlexBackend.Infrastructure.csproj
@@ -10,4 +10,8 @@
       <PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" Version="5.0.11" />
     </ItemGroup>
 
+    <ItemGroup>
+      <ProjectReference Include="..\PlexBackend.Core\PlexBackend.Core.csproj" />
+    </ItemGroup>
+
 </Project>
