+++ src/PlexBackend.Infrastructure/Repositories/Base/Repository.cs
@@ -1,11 +1,11 @@
-﻿using PlexBackend.Infrastructure;
+﻿using PlexBackend.Core.Interfaces;
+using PlexBackend.Infrastructure;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Linq.Expressions;
 using System.Text;
 using System.Threading.Tasks;
-using PlexBackend.Infrastructure.Interfaces;
 
 namespace PlexBackend.Infrastructure.Repositories.Base
 {
