+++ src/PlexBackend.WebApi/ViewModels/StudentChoiceViewModel.cs
@@ -1,5 +1,4 @@
 ﻿using AutoMapper.Configuration.Conventions;
-using PlexBackend.Infrastructure.ContextModels;
 using System;
 using System.Collections.Generic;
 using System.Linq;
